# expression: implement vectorized evaluation for 'builtinIntervalRealSig'