<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinIntervalRealSig, for #12103


### What is changed and how it works?
```
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-VecBuiltinFunc-8         	  10000     15984 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-NonVecBuiltinFunc-8      	   5000     37097 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-VecBuiltinFunc#01-8      	  10000     19508 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-NonVecBuiltinFunc#01-8   	   3000     39998 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-VecBuiltinFunc#02-8      	   5000     25621 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalRealSig-NonVecBuiltinFunc#02-8   	   3000     46326 ns/op	       0 B/op	       0 allocs/op
```
### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test